Fixing build/merge issues
