+++ pkg/BUILD.bazel
@@ -403,12 +403,15 @@ cc_dist_library(
         "//src/google/protobuf/compiler:command_line_interface",
         "//src/google/protobuf/compiler/cpp",
         "//src/google/protobuf/compiler/cpp:names",
+        "//src/google/protobuf/compiler/cpp:names_internal",
         "//src/google/protobuf/compiler/csharp",
         "//src/google/protobuf/compiler/csharp:names",
         "//src/google/protobuf/compiler/java",
         "//src/google/protobuf/compiler/java:names",
+        "//src/google/protobuf/compiler/java:names_internal",
         "//src/google/protobuf/compiler/objectivec",
         "//src/google/protobuf/compiler/objectivec:names",
+        "//src/google/protobuf/compiler/objectivec:names_internal",
         "//src/google/protobuf/compiler/php",
         "//src/google/protobuf/compiler/php:names",
         "//src/google/protobuf/compiler/python",
