+++ src/google/protobuf/compiler/BUILD.bazel
@@ -243,6 +243,7 @@ cc_test(
         ":command_line_interface",
         ":mock_code_generator",
         "//:protobuf",
+        "//src/google/protobuf/compiler/cpp:names",
         "//src/google/protobuf:cc_test_protos",
         "//src/google/protobuf:test_util2",
         "//src/google/protobuf/io",
