+++ src/google/protobuf/compiler/cpp/BUILD.bazel
@@ -10,16 +10,33 @@ load("//build_defs:cpp_opts.bzl", "COPTS")
 cc_library(
     name = "names",
     hdrs = ["names.h"],
-    srcs = [
-        # Internal dependencies.
-        "helpers.h",
-        "helpers.cc",
-        "options.h",
-    ],
     copts = COPTS,
     include_prefix = "google/protobuf/compiler/cpp",
     visibility = ["//visibility:public"],
-    deps = ["//src/google/protobuf:protobuf_nowkt"],
+    deps = [
+        ":names_internal",
+        "//src/google/protobuf/compiler:code_generator",
+        "//src/google/protobuf:protobuf_nowkt",
+    ],
+)
+
+cc_library(
+    name = "names_internal",
+    hdrs = [
+        "helpers.h",
+        "names.h",
+        "options.h",
+    ],
+    srcs = [
+        "helpers.cc",
+    ],
+    copts = COPTS,
+    include_prefix = "google/protobuf/compiler/cpp",
+    visibility = ["//pkg:__pkg__"],
+    deps = [
+        "//src/google/protobuf/compiler:code_generator",
+        "//src/google/protobuf:protobuf_nowkt",
+    ],
 )
 
 cc_library(
@@ -64,6 +81,7 @@ cc_library(
         "//src/google/protobuf/compiler:__pkg__",
     ],
     deps = [
+        ":names_internal",
         ":names",
         "//src/google/protobuf:protobuf_nowkt",
         "//src/google/protobuf/compiler:code_generator",
