+++ src/google/protobuf/compiler/java/BUILD.bazel
@@ -9,17 +9,33 @@ load("//build_defs:cpp_opts.bzl", "COPTS")
 cc_library(
     name = "names",
     hdrs = ["names.h"],
-    srcs = [
-        # Internal dependencies.
-        "names.cc",
-        "name_resolver.cc",
-        "name_resolver.h",
-        "options.h",
-    ],
     copts = COPTS,
     include_prefix = "google/protobuf/compiler/java",
     visibility = ["//visibility:public"],
     deps = [
+        ":names_internal",
+        "//src/google/protobuf:protobuf_nowkt",
+    ],
+)
+
+cc_library(
+    name = "names_internal",
+    hdrs = [
+        "helpers.h",
+        "name_resolver.h",
+        "names.h",
+        "options.h",
+    ],
+    srcs = [
+        "helpers.cc",
+        "name_resolver.cc",
+        "names.cc",
+    ],
+    copts = COPTS,
+    include_prefix = "google/protobuf/compiler/java",
+    visibility = ["//pkg:__pkg__"],
+    deps = [
+        "//src/google/protobuf/compiler:code_generator",
         "//src/google/protobuf:protobuf_nowkt",
         "@com_google_absl//absl/container:flat_hash_set",
     ],
@@ -40,7 +56,6 @@ cc_library(
         "file.cc",
         "generator.cc",
         "generator_factory.cc",
-        "helpers.cc",
         "kotlin_generator.cc",
         "map_field.cc",
         "map_field_lite.cc",
@@ -71,7 +86,6 @@ cc_library(
         "file.h",
         "generator.h",
         "generator_factory.h",
-        "helpers.h",
         "kotlin_generator.h",
         "map_field.h",
         "map_field_lite.h",
@@ -96,6 +110,7 @@ cc_library(
         "//src/google/protobuf/compiler:__pkg__",
     ],
     deps = [
+        ":names_internal",
         ":names",
         "//src/google/protobuf:protobuf_nowkt",
         "//src/google/protobuf/compiler:code_generator",
