+++ src/google/protobuf/compiler/java/names.h
@@ -40,6 +40,7 @@
 
 #include <string>
 
+#include "google/protobuf/descriptor.h"
 #include "google/protobuf/compiler/java/options.h"
 
 namespace google {
