+++ src/google/protobuf/compiler/php/names.cc
@@ -28,7 +28,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "google/protobuf/compiler/php/php_generator.h"
+#include "google/protobuf/compiler/php/names.h"
 
 #include <string>
